+++ server/shared/dotnet/caching/abstractions/InputFailures.cs
@@ -38,4 +38,24 @@ public static class InputFailures
     public static D2Result Required(string paramName)
         => D2Result.ValidationFailed(
             inputErrors: [new InputError(paramName, [TK.Common.Errors.NOT_NULL_VIOLATION])]);
+
+    /// <summary>
+    /// Builds a <see cref="D2Result{TData}"/> input failure for a present but
+    /// invalid parameter value (range, non-positive, etc.). Does not use
+    /// <see cref="TK.Common.Errors.NOT_NULL_VIOLATION"/>.
+    /// </summary>
+    /// <typeparam name="T">Wrapped result type.</typeparam>
+    /// <param name="paramName">Parameter name (use <c>nameof</c>).</param>
+    public static D2Result<T> Invalid<T>(string paramName)
+        => D2Result<T>.ValidationFailed(
+            inputErrors: [new InputError(paramName, [TK.Common.Errors.VALIDATION_FAILED])]);
+
+    /// <summary>
+    /// Builds a non-generic <see cref="D2Result"/> input failure for a present
+    /// but invalid parameter value (range, non-positive, etc.).
+    /// </summary>
+    /// <param name="paramName">Parameter name (use <c>nameof</c>).</param>
+    public static D2Result Invalid(string paramName)
+        => D2Result.ValidationFailed(
+            inputErrors: [new InputError(paramName, [TK.Common.Errors.VALIDATION_FAILED])]);
 }