+++ server/shared/typescript/caching/local-default/src/index.ts
@@ -3,4 +3,9 @@
 // -----------------------------------------------------------------------
 
 export { DefaultLocalCache } from "./default-local-cache.js";
-export { LOCAL_CACHE_METER_NAME } from "./local-cache-telemetry.js";
+export {
+  LOCAL_CACHE_INSTRUMENTS,
+  LOCAL_CACHE_METER_NAME,
+  LOCAL_CACHE_METER_VERSION,
+  type CacheInstrumentMeta,
+} from "./local-cache-telemetry.js";